+++ packages/network/bin/tracker.js
@@ -9,7 +9,6 @@ const { MetricsContext } = require('../dist/src/helpers/MetricsContext')
 program
     .version(CURRENT_VERSION)
     .option('--id <id>', 'Ethereum address / tracker id', undefined)
-    .option('--trackerName <trackerName>', 'Human readable name', undefined)
     .option('--port <port>', 'port', '27777')
     .option('--ip <ip>', 'ip', '0.0.0.0')
     .option('--unixSocket <unixSocket>', 'unixSocket', undefined)
@@ -22,7 +21,6 @@ program
     .parse(process.argv)
 
 const id = program.opts().id || `TR${program.opts().port}`
-const name = program.opts().trackerName || id
 const logger = new Logger(module)
 const listen = program.opts().unixSocket ? program.opts().unixSocket : {
     hostname: program.opts().ip,
@@ -43,12 +41,11 @@ const getTopologyStabilization = () => {
 }
 
 async function main() {
-    const metricsContext = new MetricsContext(id)
+    const metricsContext = new MetricsContext()
     try {
         await startTracker({
             listen,
             id,
-            name,
             maxNeighborsPerNode: Number.parseInt(program.opts().maxNeighborsPerNode, 10),
             metricsContext,
             topologyStabilization: getTopologyStabilization()
@@ -62,7 +59,6 @@ async function main() {
 
         logger.info('started tracker: %o', {
             id,
-            name,
             ...trackerObj
         })