+++ packages/client/CHANGELOG.md
@@ -37,9 +37,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove (non-functional) client configuration options `autoConnect` and `autoDisconnect`
 - Remove method `disconnect()`, use `destroy()` instead
 - Remove method `unsubscribeAll()`, use `unsubscribe()` without arguments instead
+- Remove client configuration option `client.network.name`
 
 ### Fixed
 
+- Fix stream encryption: messages weren't automatically encrypted if the local database didn't contain pre-existing encryption keys for a stream
 - Fix timeout issue of method `addToStorageNode` when used with storage node cluster
 
 ### Security